AuditLoggerEvent blir brukt frå mange stadar av anna kode, så denne blir jo brukt. Så trekkjer ut denne til si eiga fil og har denne framleis liggjande i config-pakka
