+++ src/main/kotlin/no/nav/familie/ba/sak/config/AuditLoggerEvent.kt
@@ -1,0 +1,8 @@
+package no.nav.familie.ba.sak.config
+
+enum class AuditLoggerEvent(val type: String) {
+    CREATE("create"),
+    UPDATE("update"),
+    DELETE("delete"),
+    ACCESS("access")
+}
